refactor: split policy engine kernel
