+++ crates/assay-core/src/mcp/policy/engine_next/mod.rs
@@ -1,0 +1,17 @@
+mod diagnostics;
+mod effects;
+mod fail_closed;
+mod matcher;
+mod precedence;
+
+#[cfg(test)]
+pub(in crate::mcp::policy) use diagnostics::parse_delegation_context;
+pub(in crate::mcp::policy) use diagnostics::{apply_delegation_context, finalize_evaluation};
+pub(in crate::mcp::policy) use effects::{
+    apply_approval_required_obligation, apply_redact_args_obligation,
+    apply_restrict_scope_obligation,
+};
+pub(in crate::mcp::policy) use fail_closed::{
+    check_rate_limits, schema_violation_decision, tool_drift_decision, unconstrained_decision,
+};
+pub(in crate::mcp::policy) use precedence::{apply_allow_precedence, deny_match_decision};
