+++ crates/assay-core/src/mcp/policy/mod.rs
@@ -1,4 +1,5 @@
 mod engine;
+mod engine_next;
 mod legacy;
 mod response;
 mod schema;
